Refactor Entity Explosion
Pulls out EntityExplosion into a common file.

`cen` and `rwrp` appear to have not imported a declaration for
`AnimateEntity` and call it as an undeclared function with a int-sized
return value. Since AnimateEntity is declared in the same compilation
unit a workaround is done to fix how the return value is interpreted.
