+++ config/symbols.pspeu.stwrp.txt
@@ -77,8 +77,8 @@ g_ESoulStealOrbAnim = 0x09246138;
 g_goldCollectTexts = 0x092464A8;
 c_GoldPrizes = 0x092464D0;
 c_HeartPrizes = 0x09246500;
-D_80180EC4 = 0x09246508;
-D_80180F70 = 0x092465C8;
+g_ExplosionYVelocities = 0x09246508;
+g_ExplosionAnimations = 0x092465C8;
 unk15_rot = 0x09246690;
 unk15_yVel = 0x092466A0;
 unk14_yVel = 0x092466B8;
