+++ config/symbols.us.stcen.txt
@@ -22,7 +22,9 @@ g_eRedDoorUV = 0x80180BA4;
 g_goldCollectTexts = 0x80180F30;
 c_GoldPrizes = 0x80180F58;
 c_HeartPrizes = 0x80180FE8;
+g_ExplosionYVelocities = 0x80180FEC;
 g_bigRedFireballAnim = 0x80181014;
+g_ExplosionAnimations = 0x80181098;
 g_eUnk14SpawnRots = 0x80181110;
 unk15_rot = 0x80181120;
 unk15_yVel = 0x80181130;
