+++ config/symbols.us.stdre.txt
@@ -29,7 +29,9 @@ g_eRedDoorUV = 0x80180D68;
 g_goldCollectTexts = 0x801810F4;
 c_GoldPrizes = 0x8018111C;
 c_HeartPrizes = 0x801811AC;
+g_ExplosionYVelocities = 0x801811B0;
 g_bigRedFireballAnim = 0x801811D8;
+g_ExplosionAnimations = 0x8018125C;
 g_eUnk14SpawnRots = 0x801812D4;
 unk15_rot = 0x801812E4;
 unk15_yVel = 0x801812F4;
