+++ config/symbols.us.stno3.txt
@@ -29,7 +29,9 @@ g_eRedDoorUV = 0x801820AC;
 g_goldCollectTexts = 0x80182438;
 c_GoldPrizes = 0x80182460;
 c_HeartPrizes = 0x801824F0;
+g_ExplosionYVelocities = 0x801824F4;
 g_bigRedFireballAnim = 0x8018251C;
+g_ExplosionAnimations = 0x801825A0;
 g_eUnk14SpawnRots = 0x80182618;
 unk15_rot = 0x80182628;
 unk15_yVel = 0x80182638;
