+++ config/symbols.us.stnz0.txt
@@ -29,7 +29,9 @@ g_eRedDoorUV = 0x80181960;
 g_goldCollectTexts = 0x80181CEC;
 c_GoldPrizes = 0x80181D14;
 c_HeartPrizes = 0x80181DA4;
+g_ExplosionYVelocities = 0x80181DA8;
 g_bigRedFireballAnim = 0x80181DD0;
+g_ExplosionAnimations = 0x80181E54;
 g_eUnk14SpawnRots = 0x80181ECC;
 unk15_rot = 0x80181EDC;
 unk15_yVel = 0x80181EEC;
