+++ config/symbols.us.strwrp.txt
@@ -17,7 +17,9 @@ g_testCollRandTable = 0x80180A50;
 g_goldCollectTexts = 0x80180E08;
 c_GoldPrizes = 0x80180E30;
 c_HeartPrizes = 0x80180EC0;
+g_ExplosionYVelocities = 0x80180EC4;
 g_bigRedFireballAnim = 0x80180EEC;
+g_ExplosionAnimations = 0x80180F70;
 g_eUnk14SpawnRots = 0x80180FE8;
 unk15_rot = 0x80180FF8;
 unk15_yVel = 0x80181008;
