+++ config/symbols.us.stst0.txt
@@ -18,7 +18,9 @@ g_testColliFrames = 0x8018195C;
 g_eDamageDisplayClut = 0x80181968;
 g_eRedDoorUV = 0x80181978;
 c_GoldPrizes = 0x80181CF0;
+g_ExplosionYVelocities = 0x80181D7C;
 g_bigRedFireballAnim = 0x80181DA4;
+g_ExplosionAnimations = 0x80181E28;
 g_eUnk14SpawnRots = 0x80181EA0;
 unk15_rot = 0x80181EB0;
 unk15_yVel = 0x80181EC0;
