+++ src/st/nz0/3E30C.c
@@ -217,36 +217,7 @@ void EntityPrizeDrop(Entity* self) {
     }
 }
 
-void EntityExplosion(Entity* entity) {
-    u32 temp_v0;
-
-    if (entity->step == 0) {
-        InitializeEntity(g_InitializeEntityData0);
-        entity->animSet = ANIMSET_DRA(2);
-        entity->animFrameIdx = 0;
-        entity->animFrameDuration = 0;
-        entity->drawMode = 0x30;
-
-        if (entity->params & 0xF0) {
-            entity->palette = 0x8195;
-            entity->drawMode = DRAW_TPAGE;
-        }
-
-        temp_v0 = entity->params & 0xFF00;
-        if (temp_v0) {
-            entity->zPriority = temp_v0 >> 8;
-        }
-
-        entity->params = entity->params & 0xF;
-        entity->velocityY = D_80181DA8[entity->params];
-        return;
-    }
-
-    entity->posY.val += entity->velocityY;
-    if (!AnimateEntity(D_80181E54[entity->params], entity)) {
-        DestroyEntity(entity);
-    }
-}
+#include "../entity_explosion.h"
 
 #include "../blink_item.h"
