+++ src/st/nz0/nz0.h
@@ -49,7 +49,7 @@ void func_801C4CC0(void);
 
 extern u8 D_8003BE6F[];
 extern PfnEntityUpdate PfnEntityUpdates[];
-extern const u16 g_InitializeEntityData0[];
+extern u16 g_InitializeEntityData0[];
 extern u16 D_80180BEC[];
 extern u16 g_EInitGeneric[]; // Init Elevator2
 extern u16 g_eInitGeneric2[];
