+++ src/st/st0/st0.h
@@ -158,8 +158,6 @@ extern s32 D_801C2578;
 extern s16 D_80181990[];
 extern u16 g_InitializeData0[];
 extern u16 D_80181D74[];
-extern u32 D_80181D7C[];
-extern u8* D_80181E28[];
 extern s16 D_80181E3C[];
 extern s16 D_80181E3E[];
 extern s16 D_80181EB0[];
