+++ src/st/wrp/e_collect.c
@@ -335,35 +335,8 @@ void EntityPrizeDrop(Entity* self) {
 }
 
 extern u16 g_InitializeEntityData0[];
-extern s32 D_80180EC4[];
-extern u8* D_80180F70[];
-void EntityExplosion(Entity* entity) {
-    u16 zPriority;
 
-    if (entity->step == 0) {
-        InitializeEntity(g_InitializeEntityData0);
-        entity->animSet = ANIMSET_DRA(2);
-        entity->animFrameIdx = 0;
-        entity->animFrameDuration = 0;
-        entity->drawMode = 0x30;
-        if (entity->params & 0xF0) {
-            entity->palette = 0x8195;
-            entity->drawMode = DRAW_TPAGE;
-        }
-
-        zPriority = entity->params & 0xFF00;
-        if (zPriority) {
-            entity->zPriority = zPriority >> 8;
-        }
-        entity->params &= 15;
-        entity->velocityY = D_80180EC4[entity->params];
-    } else {
-        entity->posY.val += entity->velocityY;
-        if (!AnimateEntity(D_80180F70[entity->params], entity)) {
-            DestroyEntity(entity);
-        }
-    }
-}
+#include "../entity_explosion.h"
 
 #include "../blink_item.h"
