+++ src/st/wrp_psp/e_collect.c
@@ -132,33 +132,7 @@ INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_collect", DestroyCurrentEntity);
 
 INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_collect", EntityPrizeDrop);
 
-extern u16 g_InitializeEntityData0[];
-extern s32 D_80180EC4[];
-extern u8* D_80180F70[];
-void EntityExplosion(Entity* entity) {
-    if (!entity->step) {
-        InitializeEntity(g_InitializeEntityData0);
-        entity->animFrameIdx = 0;
-        entity->animFrameDuration = 0;
-        entity->animSet = ANIMSET_DRA(2);
-        entity->drawMode = 0x30;
-        if (entity->params & 0xF0) {
-            entity->palette = 0x8195;
-            entity->drawMode = DRAW_TPAGE;
-        }
-
-        if (entity->params & 0xFF00) {
-            entity->zPriority = (entity->params & 0xFF00) >> 8;
-        }
-        entity->params &= 15;
-        entity->velocityY = D_80180EC4[entity->params];
-    } else {
-        entity->posY.val += entity->velocityY;
-        if (!AnimateEntity(D_80180F70[entity->params], entity)) {
-            DestroyEntity(entity);
-        }
-    }
-}
+#include "../entity_explosion.h"
 
 #include "../blink_item.h"
